[core] - fix: Device Check (#1034)
* Refactor device checks

* Fix build order

* Increment version

* Remove flag for dev
