+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@web3-onboard/core",
-  "version": "2.2.9",
+  "version": "2.2.9-alpha.2",
   "scripts": {
     "build": "rollup -c",
     "dev": "rollup -c -w",
