+++ packages/core/src/internals.ts
@@ -1,0 +1,8 @@
+import type { InternalState } from './types'
+import { getDevice } from './utils'
+
+export const internalState: InternalState = {
+  svelteInstance: null,
+  appMetadata: null,
+  device: getDevice()
+}
