+++ packages/core/src/store/actions.ts
@@ -1,6 +1,4 @@
-import type { Chain, WalletInit } from '@web3-onboard/common'
-import { internalState$ } from '../streams'
-import { initializeWalletModules, uniqueWalletsByLabel } from '../utils'
+import type { Chain, WalletInit, WalletModule } from '@web3-onboard/common'
 import { dispatch } from './index'
 
 import type {
@@ -37,6 +35,7 @@ import {
   SET_WALLET_MODULES,
   SET_LOCALE
 } from './constants'
+import { internalState } from '../internals'
 
 export function addChains(chains: Chain[]): void {
   // chains are validated on init
@@ -152,10 +151,7 @@ export function setWalletModules(wallets: WalletInit[]): void {
     throw error
   }
 
-  const modules = initializeWalletModules(
-    wallets,
-    internalState$.getValue().device
-  )
+  const modules = initializeWalletModules(wallets)
   const dedupedWallets = uniqueWalletsByLabel(modules)
 
   const action = {
@@ -179,4 +175,30 @@ export function setLocale(locale: string): void {
   }
 
   dispatch(action as SetLocaleAction)
+}
+
+// ==== HELPERS ==== //
+export function initializeWalletModules(modules: WalletInit[]): WalletModule[] {
+  const { device } = internalState
+  return modules.reduce((acc, walletInit) => {
+    const initialized = walletInit({ device })
+
+    if (initialized) {
+      // injected wallets is an array of wallets
+      acc.push(...(Array.isArray(initialized) ? initialized : [initialized]))
+    }
+
+    return acc
+  }, [] as WalletModule[])
+}
+
+export function uniqueWalletsByLabel(
+  walletModuleList: WalletModule[]
+): WalletModule[] {
+  return walletModuleList.filter(
+    (wallet, i) =>
+      walletModuleList.findIndex(
+        (innerWallet: WalletModule) => innerWallet.label === wallet.label
+      ) === i
+  )
 }
