+++ packages/core/src/views/account-center/Minimized.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { fade } from 'svelte/transition'
-  import { internalState$, wallets$ } from '../../streams'
+  import { wallets$ } from '../../streams'
   import {
     getDefaultChainStyles,
     shortenAddress,
@@ -15,8 +15,9 @@
   import WalletAppBadge from '../shared/WalletAppBadge.svelte'
   import NetworkSelector from '../shared/NetworkSelector.svelte'
   import { state } from '../../store'
+  import { internalState } from '../../internals'
 
-  const { appMetadata } = internalState$.getValue()
+  const { appMetadata } = internalState
   const appIcon = (appMetadata && appMetadata.icon) || questionIcon
   const chains = state.get().chains
