+++ packages/core/src/views/connect/Sidebar.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import { _ } from 'svelte-i18n'
   import blocknative from '../../icons/blocknative'
-  import { internalState$ } from '../../streams'
   import en from '../../i18n/en.json'
   import type { i18n } from '../../types'
   import { isSVG } from '../../utils'
+  import { internalState } from '../../internals'
 
   export let step: keyof i18n['connect']
 
-  const { appMetadata } = internalState$.getValue()
+  const { appMetadata } = internalState
   const { icon, logo, name = 'This app' } = appMetadata || {}
 
   const defaultContent = en.connect[step].sidebar
